0.0.1

+++ contracts/CheapMintNFT.sol
@@ -81,7 +81,7 @@ contract CheapMintNFT is ICheapMintNFT {
         uint256 tokenId,
         bytes calldata
     ) external pure override returns (bytes4) {
-        revert(string(abi.encode(tokenId)));
+        revert(string(abi.encode(uint32(tokenId))));
     }
 }
